% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hugo.r
\name{hugo}
\alias{hugo}
\title{Convert to markdown for a Hugo blog post}
\usage{
hugo(toc = FALSE, toc_depth = 3, fig_width = 7, fig_height = 5,
  fig_retina = 2, dev = "png", includes = NULL,
  md_extensions = NULL, pandoc_args = NULL)
}
\arguments{
\item{toc}{\code{TRUE} to include a table of contents in the output}

\item{toc_depth}{Depth of headers to include in table of contents}

\item{fig_width}{Default width (in inches) for figures}

\item{fig_height}{Default height (in inches) for figures}

\item{fig_retina}{Scaling to perform for retina displays. Defaults to
\code{NULL} which performs no scaling. A setting of 2 will work for all
widely used retina displays, but will also result in the output of
\code{<img>} tags rather than markdown images due to the need to set the
width of the image explicitly.}

\item{dev}{Graphics device to use for figure output (defaults to png)}

\item{includes}{Named list of additional content to include within the
document (typically created using the \code{\link{includes}} function).}

\item{md_extensions}{Markdown extensions to be added or removed from the
default definition or R Markdown. See the \code{\link{rmarkdown_format}} for
additional details.}

\item{pandoc_args}{Additional command line options to pass to pandoc}
}
\value{
R Markdown output format to pass to \code{\link{render}}
}
\description{
Format for converting from R Markdown to markdown for Hugo
}
