% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdowntemplates-package.R
\docType{package}
\name{markdowntemplates}
\alias{markdowntemplates}
\alias{markdowntemplates-package}
\title{Alternate R markdown templates}
\description{
These are a set of alternate R markdown templates that do not use Bootstrap
and do not include or rely on jQuery. These are primarily here for
demonstration purposes but are also useful in and of themselves (it's always
good to have choices). If you have a particular framework you like and it's
not here, file an issue and I'll add it.
}
\details{
`bulma` (`output: markdowntemplates::bulma`) uses the [Bulma CSS
framework](http://bulma.io) and has optional `navtitle` and `navlink` YAML
header elements which will be used in header of the generated file.

`skeleton` (`output: markdowntemplates::skeleton`) or `default` (`output:
markdowntemplates::default`) uses the [Skeleton CSS
framework](http://getskeleton.com) and also has optional `navtitle` and
`navlink` YAML header elements which will be used in header of the generated
file.

`minimal` (`output: markdowntemplates::minimal`) uses no CSS framework and
only uses the `title` YAML header element.

When you use RStudio to create a new R Markdown document, select "From
Template" and choose one of these templates.
}
\author{
Bob Rudis (@hrbrmstr)
}

