% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-prismpress.R
\name{prismpress_source_hook}
\alias{prismpress_source_hook}
\alias{prism_press_chunk_hook}
\title{Custom knitr hook for prismpress}
\usage{
prismpress_source_hook(x, options)

prism_press_chunk_hook(x, options)
}
\arguments{
\item{x, options}{knitr hook option}
}
\description{
Custom knitr hook for prismpress
}
