% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdowntemplates-package.R
\docType{package}
\name{markdowntemplates}
\alias{markdowntemplates}
\alias{markdowntemplates-package}
\title{Personal R markdown templates}
\description{
Personal R markdown templates
}
\author{
Bob Rudis (@hrbrmstr)
}

